feat: cart basic tests
